Wrap returns of `errno` from `main()` with `EXIT_RC()`.
Returns of error code from `main()` prevent type checking when `errno_t`
and `int` are considered incompatible. In order to avoid the philosophical
discussion of what should and shouldn't be returned for `main()`, we simply
wrap the error values and leave the answer to the question for future
generations to decide.