Expect cross toolchain binaries to be in PATH
The cross files for Meson are written with this expectation, since
it allows us to not have to generate those files for custom user
paths. It's also convenient for general cross-building of code
to HelenOS (one can simply do CC=amd64-helenos-gcc to cross build
simple projects).

Previously, the Makefile set the PATH to preserve old user experience.
However that confusingly breaks when user tries to run ninja manually,
so we move the responsibility for PATH entirely to the user.