Modify semantics of WAKEUP_ALL slightly
Previous implementation saves the wakeup if and only if no threads
have been woken up. IMO that is a counterintuitive behaviour and
limits usefulness of the function.

For context, the only current consumer of WAKEUP_ALL is condvar
broadcast, in which case there is always one extra spurious
wakeup going out. Condvar allows spurious wakeups, so it's
not an issue here, but there are other possible uses where
it can be a problem.