Replace context_save/restore with context_create/swap
Similar to the past switch in libc, context_swap() allows
context switching to be done without distrupting natural
flow of the code.

We also split scheduler() into separate functions. There
are two distinct cases for entry to scheduler.
One is to switch from one thread to another, and the other
is a CPU without a thread starting to schedule threads.

The distinction will make sense in the next commit where we
make scheduler context persistent.