Make scheduler cyclical
With this commit, we make scheduler a loop that does not return,
and instead waits for return from context_swap(). With this,
we can avoid recreating the context every time. While this means
we need to store the context, it will allow the scheduler code to
flow more naturally. The extra storage needed by the context
will be mitigated since we now only use context_swap(), which
frees us to reduce context to the bare minimum.